rework PUCCH codec with c16_t type, rewrite large parts of functions to make readable code. Not fixed parts that are using SIMD. No change, the code is still same block by block (variable visibility narrowed, declaration can be shifted)

+++ openair1/PHY/NR_TRANSPORT/nr_transport_common_proto.h
@@ -42,7 +42,7 @@
 
 void nr_group_sequence_hopping(pucch_GroupHopping_t PUCCH_GroupHopping,
                                uint32_t n_id,
-                               uint8_t n_hop,
+                               int n_hop,
                                int nr_slot_tx,
                                uint8_t *u,
                                uint8_t *v);
